Esquematizando layout da lista de locais

+++ lib/models/weather_model.dart
@@ -7,7 +7,7 @@ class Weather {
   int utcOffsetSeconds;
   String timezone;
   String timezoneAbbreviation;
-  int elevation;
+  double elevation;
   HourlyUnits hourlyUnits;
   Hourly hourly;
   DailyUnits dailyUnits;
